requirements.txt
streamlit
unstructured
transformers
torch
sentence-transformers
langchain
faiss-cpu
Pillow
google-generativeai
accelerate
bitsandbytes
langchain-community
